% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{acf_computation}
\alias{acf_computation}
\title{acf_computation}
\usage{
acf_computation(csv, csv_complete)
}
\arguments{
\item{csv}{A single column of the csv_complete (given by apply())}

\item{csv_complete}{rolling_mean table output of \code{\link{format_csv}}.}
}
\description{
Function to determine where the pulse occurs in the traces (internal function of \code{\link{Finding_pics}})
}
\examples{
csv_df = format_csv(csv)

### Load csv and modify it ###
csv = csv_df[csv_df$odorant == "water 1",]
csv = csv[,-c(ncol(csv))]
csv[,-c(1,ncol(csv))] = as.data.frame(scale(csv[,-c(1,ncol(csv))],scale = F))
csv$Timing = rank(csv$Timing)

### compute acf ###
computed_pics = apply(csv[,2:(ncol(csv)-2)], 2, function(x)as.data.frame(acf_computation(x,csv)))
pulse_csv = reduce(computed_pics,left_join,by="Timing")

}
