% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_pdf}
\alias{create_pdf}
\title{create_pdf}
\usage{
create_pdf(csv_DF, odorant, z_score = F, name_file)
}
\arguments{
\item{csv_DF}{table output of \code{\link{format_csv}}.}

\item{odorant}{Corresponding odorant to generate the plots (Should be present in the exp_info.csv file).}

\item{z_score}{Default = True, plot the trace on a raw scale (False) or z score scale (True).}

\item{name_file}{File name of the PDF.}
}
\description{
Pipeline to generate a single PDF containing all plots doable with this package for 1 odor.
}
\examples{
exp.info = Read_Exp_info()
csv_split = Split_CSV(getwd())
csv_DF = format_csv(csv=csv_split)
create_pdf(csv_DF,exp.info$odorant[1],z_score = T,name_file="Odorant_1.pdf")

}
