% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{format_csv}
\alias{format_csv}
\title{format_csv}
\usage{
format_csv(csv_input, delta_F_dan = F)
}
\arguments{
\item{csv_input}{table output of \code{\link{Split_CSV}}}
}
\description{
normalize the dataset produced by Split_CSV
}
\examples{
csv = Split_csv(getwd())
df = format_csv(csv)

}
