% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LumarAnalysisIrlab-package.R
\docType{package}
\name{LumarAnalysisIrlab-package}
\alias{LumarAnalysisIrlab}
\alias{LumarAnalysisIrlab-package}
\title{LumarAnalysisIrlab: Package to analyze fluorescent data}
\description{
Analysing fluorescent data comming from a lumar microscope. Different parameters could be apply. Packages speciallized in odorant testing response of olfactory receptors.
}
\keyword{internal}
