% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pipeline}
\alias{pipeline}
\title{pipeline}
\usage{
pipeline(z_score = F)
}
\arguments{
\item{z_score}{Default = False, plot the trace on a raw scale (False) or z score scale (True).}
}
\description{
Pipeline to generate the complete analysis
}
\examples{
pipeline(z_score=T)

}
