% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_before_cutting}
\alias{plot_before_cutting}
\title{plot_before_cutting}
\usage{
plot_before_cutting(raw = NULL)
}
\arguments{
\item{raw}{raw table output of \code{\link{format_csv}}.}

\item{alone}{Default = False, To make the plot alone without running all the code}
}
\description{
Plot of raw traces before do anything to compare with results
}
\examples{
csv_split = Split_CSV(getwd())
csv_DF = format_csv(csv=csv_split)
plot_before_cutting(csv_DF$raw)

}
