% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PlotResume}
\alias{PlotResume}
\title{PlotResume}
\usage{
PlotResume(csv, groupby = "antenna", z_score = T)
}
\arguments{
\item{csv}{table output of \code{\link{Z_score_calculation}}.}

\item{z_score}{Default = True, plot the trace on a raw scale (False) or z score scale (True).}

\item{groupBy}{Default = "antenna", separate each antenna wrapped by conditions ("antenna"), If another string than antenna is given the global mean of each condition will be displayed.}
}
\description{
Resume plot with boxplots representing max values of each pulse for each antenna or condition
}
\examples{
pics = Finding_pics(csv$rolling_mean, exp_odorant="water 1")
pics$pulse_csv = z_score_calculation(pics$pulse_csv)
PlotResume(pics,groupBy="antenna",z_score=T)

}
