% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PlotTraceCondition}
\alias{PlotTraceCondition}
\title{PlotTraceCondition}
\usage{
PlotTraceCondition(csv, groupBy = "pulse", z_score = T)
}
\arguments{
\item{csv}{table output of \code{\link{Z_score_calculation}}.}

\item{groupBy}{Default = "pulse", color by pulse the overlapped traces ("pulse") or by antenna ID ("antenna").}

\item{z_score}{Default = True, plot the trace on a raw scale (False) or z score scale (True).}
}
\description{
plot the trace of fluorescence intensity overlapped for each conditions
}
\examples{
pics = Finding_pics(csv$rolling_mean, exp_odorant="water 1")
pics$pulse_csv = z_score_calculation(pics$pulse_csv)
PlotTraceCondition(pics,groupBy="pulse",z_score=T)

}
