% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{samescale_summary}
\alias{samescale_summary}
\title{samescale_summary}
\usage{
samescale_summary(for_resume_final, without_water = F)
}
\arguments{
\item{for_resume_final}{table output of \code{\link{pipeline}} or \code{\link{create_pdf}}}

\item{without_water}{Default = False, if True will skip "water" odorants.}
}
\description{
Plot of raw traces before do anything to compare with results
}
\examples{
# All functions splited
csv_split = Split_CSV(getwd())
csv_DF = format_csv(csv=csv_split)

for_resume_final = c()
for(odor in exp_info$odorant){
  for_resume = create_pdf(csv_DF,odor,z_score=T,name_file="odorant_1.pdf")
  for_resume_final = rbind(for_resume_final,for_resume)
}
samescale_summary(for_resume_final)

# All at once
for_resume_final = pipeline(z_score=T)
samescale_summary(for_resume_final)


}
