% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Split_CSV}
\alias{Split_CSV}
\title{Split_CSV}
\usage{
Split_CSV(directory, length_exp = 184)
}
\arguments{
\item{directory}{Path to the raw csv file.}

\item{length_exp}{The duration in seconds of a trial for a single odor.}
}
\description{
add the odorant columns and change names of columns
}
\examples{
csv = Split_CSV(getwd())

}
