% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Z_score_calculation}
\alias{Z_score_calculation}
\title{Z_score_calculation}
\usage{
Z_score_calculation(pulse_csv)
}
\arguments{
\item{pulse_csv}{pulse_csv table output of \code{\link{Finding_pics}}}
}
\description{
z score calculation for each odorant
}
\examples{
pics = Finding_pics(csv$rolling_mean, exp_odorant="water 1")
z_score_calculation(pics$pulse_csv)

}
