% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LumarAnalysisIrlab-package.R
\docType{package}
\name{LumarAnalysisIrlab-package}
\alias{LumarAnalysisIrlab}
\alias{LumarAnalysisIrlab-package}
\title{LumarAnalysisIrlab: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
