% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Read_Exp_info}
\alias{Read_Exp_info}
\title{Read_Exp_info}
\usage{
Read_Exp_info(file = NULL)
}
\arguments{
\item{file}{Default = NULL, Path to the exp_info.csv file. If not specified search into the current directory.}
}
\description{
Read the exp_info.csv file
}
\examples{
exp.info = Read_Exp_info
exp.info$exp_id
}
